Added support for callbacks once a notification has disappeared.
humane-js now supports passing functions to be called once a notification has animated out. I've used the prefix matching for the transition events from Zepto.js and tweaked it so the right event should be watched for each engine. Also added the event for the js animation as well. And finally I've added an example to the index page and built a minified version with the amends included.